refactor auth handlers to package auth
handlers moved out main into /api/auth/
